Move some code from cmds/switch_root to pkg/mount; export useful funcs.
